requirements.txt
gradio==4.25.0
requests==2.31.0
funasr
modelscope==1.10.0
torch==2.1.0
torchvision==0.16.0
python-dotenv==1.0.0
decord==0.6.0
scipy==1.11.0